ai-auto-scanner/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
httpx==0.27.2
python-slugify==8.0.4
rich==13.8.0
jinja2==3.1.4
